+++ front/components/agent_builder/triggers/TriggerCard.tsx
@@ -3,7 +3,7 @@ import type { AgentBuilderTriggerType } from "@app/components/agent_builder/Agen
 import { getIcon } from "@app/components/resources/resources_icons";
 import { useAuth } from "@app/lib/auth/AuthContext";
 import { describeScheduleConfig } from "@app/lib/utils/schedule_description";
-import { normalizeWebhookIcon } from "@app/lib/webhookSource";
+import { normalizeWebhookIcon } from "@app/lib/webhook_source";
 import { assertNever } from "@app/types/shared/utils/assert_never";
 import type { WebhookSourceViewType } from "@app/types/triggers/webhooks";
 import { WEBHOOK_PRESETS } from "@app/types/triggers/webhooks";
